+++ rust-frontend/src/test/java/org/sonar/rust/parser/expressions/RangeExpressionTest.java
@@ -101,35 +101,5 @@ public class RangeExpressionTest {
         ;
     }
 
-    @Test
-    public void testParsing(){
 
-        //Print out Ast node content for debugging purpose
-
-        ParserAdapter<LexerlessGrammar> parser = new ParserAdapter<>(StandardCharsets.UTF_8,RustGrammar.create().build());
-        AstNode rootNode = parser.parse("let a = 1..5;");
-        System.out.printf("[DEBUG]" + rootNode.getType());
-        Assertions.assertThat(rootNode.getType()).isSameAs(RustGrammar.COMPILATION_UNIT);
-        AstNode astNode = rootNode;
-        Assertions.assertThat(astNode.getNumberOfChildren()).isEqualTo(4);
-        System.out.println("[DEBUG]" + astNode.getType());
-        for (AstNode n : astNode.getChildren()){
-            System.out.println("[DEBUG] child" + n.getType());
-        }
-        AstNode st = astNode.getChildren().get(1);
-        System.out.println("[ELG] st is " + st);
-        for (AstNode n : st.getChildren()){
-            System.out.println("[L1]" + n.getType());
-            for (AstNode n2 : n.getChildren()){
-                System.out.println("   [L2]" + n2.getType());
-                for (AstNode n3 : n2.getChildren()){
-                    System.out.println("      [L3]" + n3.getType());
-                    for (AstNode n4 : n3.getChildren()){
-                        System.out.println("         [L4]" + n4.getType());
-                    }
-                }
-
-            }
-        }
-    }
 }
